requirement.txt
absl-py==1.0.0
alembic==1.7.5
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-sts==3.1.0
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.4
cached-property==1.5.2
cachelib==0.5.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
Deprecated==1.2.13
dominate==2.6.0
Flask==1.1.2
flask-blueprint==1.3.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
image-classifier==0.0.1
imageio==2.16.0
imgviz==1.4.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelme==4.6.0
libclang==12.0.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.2.2
mlxtend==0.19.0
netron==5.4.1
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
QtPy==2.0.1
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
service-streamer==0.1.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.29
tb-nightly==2.8.0a20211209
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchsummary==1.5.1
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
unet==0.7.7
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
